requirements.txt
annotated-types==0.7.0
ll==1.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
simple-term-menu==1.6.4
six==1.16.0
typing_extensions==4.12.2
